refactor(utils): Move database & jobs utilities to subdirectories
Phase 1 (Day 3): Low-risk migration completion
- Moved database utilities to src/utils/database/ (1 file)
- Moved jobs utilities to src/utils/jobs/ (2 files)
- Fixed relative import paths in csv-processor-core.ts
- Updated 3 source imports + 1 test import
- Tests passing (199/199 smoke tests)

Files moved:
Database:
- d1-wrapper.ts

Jobs:
- progress-reporter.ts
- csv-processor-core.ts

Phase 1 complete: 9 files moved (R2, analytics, database, jobs)
